fixup! feat: transcode_music

+++ server/scripts/transcode_music/transcode.bash
@@ -1,32 +1,8 @@
 #!/bin/bash
 
-# set -x
+#set -x
 
-export MODE_DELETE=false
-export MODE_CHECKSUM=false
-
-while getopts 'fdc:' OPTION; do
-    case "$OPTION" in
-        f)
-        echo "INFO: FULL MODE"
-        export TRANSCODE_FD_FILTERS=""
-        ;;
-        d)
-        echo "INFO: DELETE MODE"
-        export MODE_DELETE=true
-        ;;
-        c)
-        echo "INFO: CHECKSUM MODE"
-        export MODE_CHECKSUM=true
-        ;;
-        ?)
-        echo "script usage: $(basename \$0) [-f] [-d] [-c]" >&2
-        exit 1
-        ;;
-    esac
-done
-shift "$(($OPTIND -1))"
-
+exec > >(awk '{ print strftime("[%Y-%m-%d %H:%M:%S]"), $0 }') 2>&1
 
 checkForVariable()
 {
@@ -38,63 +14,82 @@ checkForVariable()
     fi
 }
 
-init()
-{
+export MODE_DELETE=false
+export MODE_CHECKSUM=false
+export MODE_DRY_RUN=false
+export TIMESTAMP=$(date "+%Y%m%d_%H%M%S")
 
-    checkForVariable TRANSCODE_INPUT_DIR
-    checkForVariable TRANSCODE_OUTPUT_DIR
+checkForVariable TRANSCODE_INPUT_DIR
+checkForVariable TRANSCODE_OUTPUT_DIR
 
-    if [ -z "$TRANSCODE_DB" ]; then
-        export TRANSCODE_DB="$TRANSCODE_OUTPUT_DIR.transcode"
-    fi
-    if [ -z "$TRANSCODE_FREAC_BIN" ]; then
-        export TRANSCODE_FREAC_BIN="/app/freac.AppImage"
-    fi
-    if [ -z "$TRANSCODE_COVER_EXTENSIONS" ]; then
-        export TRANSCODE_COVER_EXTENSIONS="png jpg"
-    fi
-    if [ -z "$TRANSCODE_MUSIC_EXTENSIONS" ]; then
-        export TRANSCODE_MUSIC_EXTENSIONS="flac opus mp3 ogg wma m4a"
-    fi
-    if [ -z "$TRANSCODE_FD_FILTERS" ]; then
-        export TRANSCODE_FD_FILTERS="--changed-within 1week"
-    fi
-    
-    if [ ! -d "$TRANSCODE_INPUT_DIR" ]; then
-        echo "ERROR: Directory $TRANSCODE_INPUT_DIR does not exists"
+if [ -z "$TRANSCODE_DB" ]; then
+    export TRANSCODE_DB="$TRANSCODE_OUTPUT_DIR.transcode"
+fi
+
+if [ -z "$TRANSCODE_FREAC_BIN" ]; then
+    export TRANSCODE_FREAC_BIN="/app/freac.AppImage"
+fi
+if [ -z "$TRANSCODE_COVER_EXTENSIONS" ]; then
+    export TRANSCODE_COVER_EXTENSIONS="png jpg"
+fi
+if [ -z "$TRANSCODE_MUSIC_EXTENSIONS" ]; then
+    export TRANSCODE_MUSIC_EXTENSIONS="flac opus mp3 ogg wma m4a wav"
+fi
+if [ -z "$TRANSCODE_FD_FILTERS" ]; then
+    export TRANSCODE_FD_FILTERS="--changed-within 1week"
+fi
+
+if [ ! -d "$TRANSCODE_INPUT_DIR" ]; then
+    echo "ERROR: Directory $TRANSCODE_INPUT_DIR does not exists"
+    exit 1
+fi
+if [ ! -d "$TRANSCODE_OUTPUT_DIR" ]; then
+    echo "ERROR: Directory $TRANSCODE_OUTPUT_DIR does not exists"
+    exit 1
+fi
+if [ ! -f "$TRANSCODE_FREAC_BIN" ]; then
+    echo "ERROR: Binary $TRANSCODE_FREAC_BIN does not exists"
+    exit 1
+fi
+grep -q ID_LIKE=debian /etc/os-release
+if [ $? -eq 0 ]; then
+    export TRANSCODE_FD_BIN="fdfind"
+else
+    export TRANSCODE_FD_BIN="fd"
+fi
+
+test ! -d $TRANSCODE_DB && mkdir -p $TRANSCODE_DB
+
+if [ ! -f "$(pwd)/transcode_exclude.cfg" ]; then
+    echo "ERROR : transcode_exclude.cfg file is missing"
+    exit 1
+fi
+
+
+while getopts ':frcd' OPTION; do
+    case "$OPTION" in
+        f)
+        echo "INFO: FULL MODE"
+        export TRANSCODE_FD_FILTERS=""
+        ;;
+        r)
+        echo "INFO: DELETE MODE"
+        export MODE_DELETE=true
+        ;;
+        c)
+        echo "INFO: CHECKSUM MODE"
+        export MODE_CHECKSUM=true
+        ;;
+        d)
+        echo "INFO: DRY RUN MODE"
+        export MODE_DRY_RUN=true
+        ;;
+        ?)
+        echo "script usage: $(basename \$0) [-f] [-r] [-c] [-d]"
         exit 1
-    fi
-    if [ ! -d "$TRANSCODE_OUTPUT_DIR" ]; then
-        echo "ERROR: Directory $TRANSCODE_OUTPUT_DIR does not exists"
-        exit 1
-    fi
-    if [ ! -f "$TRANSCODE_FREAC_BIN" ]; then
-        echo "ERROR: Binary $TRANSCODE_FREAC_BIN does not exists"
-        exit 1
-    fi
-    grep -q ID_LIKE=debian /etc/os-release
-    if [ $? -eq 0 ]; then
-        export TRANSCODE_FD_BIN="fdfind"
-    else
-        export TRANSCODE_FD_BIN="fd"
-    fi
-
-    test ! -d $TRANSCODE_DB && mkdir -p $TRANSCODE_DB
-
-    if [ ! -f "$(pwd)/transcode_exclude.cfg" ]; then
-        echo "ERROR : transcode_exclude.cfg file is missing"
-        exit 1
-    fi
-
-    cp ./transcode_exclude.cfg $TRANSCODE_INPUT_DIR/.fdignore
-    cp ./transcode_exclude.cfg $TRANSCODE_OUTPUT_DIR/.fdignore
-}
-
-clean()
-{
-    rm $TRANSCODE_INPUT_DIR/.fdignore
-    rm $TRANSCODE_OUTPUT_DIR/.fdignore
-}
+        ;;
+    esac
+done
 
 transcode()
 {   
@@ -102,26 +97,33 @@ transcode()
     output_file=$2
     md5_file=$3
 
-    $TRANSCODE_FREAC_BIN freaccmd --encoder=opus --bitrate 64 "$input_file" -o "$output_file"
-    if [ $? -ne 0 ]; then exit 1; fi  
-    mkdir -p "$TRANSCODE_DB/$(dirname "$input_file")"
-    echo "$(md5sum "$input_file" | awk '{ print $1 }')" > "$md5_file"
+    if [ $MODE_DRY_RUN == false ]; then
+        $TRANSCODE_FREAC_BIN freaccmd --encoder=opus --bitrate 64 "$input_file" -o "$output_file"
+        if [ $? -ne 0 ]; then exit 1; fi  
+        mkdir -p "$TRANSCODE_DB/$(dirname "$input_file")"
+        echo "$(md5sum "$input_file" | awk '{ print $1 }')" > "$md5_file"
+    else
+        echo "INFO: transcoding $1"
+    fi
 }
 
 write_cue()
 {
     input_file=$1
-    replacement_string=$2
-    md5_file=$3
+    output_file=$2
+    replacement_string=$3
+    md5_file=$4
 
     echo "##: writing $input_file"
-    sed -i "/FILE/c $replacement_string" "$input_file"
-    if [ $? -ne 0 ]; then
-        echo "ERROR: writing cuefile $input_file"
-        exit 1
+    if [ $MODE_DRY_RUN == false ]; then
+        sed -i "/FILE/c $replacement_string" "$output_file"
+        if [ $? -ne 0 ]; then
+            echo "ERROR: writing cuefile $output_file"
+            exit 1
+        fi
+        mkdir -p "$TRANSCODE_DB/$(dirname "$input_file")"
+        echo "$(md5sum "$input_file" | awk '{ print $1 }')" > "$md5_file"
     fi
-    mkdir -p "$TRANSCODE_DB/$(dirname "$input_file")"
-    echo "$(md5sum "$input_file" | awk '{ print $1 }')" > "$md5_file"
 }
 
 write_jpg()
@@ -131,22 +133,32 @@ write_jpg()
     md5_file=$3
 
     echo "##: converting cover $input_file"
-    convert "$input_file" -resize 500 -quality 75 "$output_file"
-    if [ $? -ne 0 ]; then
-        echo "ERROR: converting cover $input_file"
-        exit 1
+    if [ $MODE_DRY_RUN == false ]; then
+        convert "$input_file" -resize 500 -quality 75 "$output_file"
+        if [ $? -ne 0 ]; then
+            echo "ERROR: converting cover $input_file"
+            exit 1
+        fi
+        mkdir -p "$TRANSCODE_DB/$(dirname "$input_file")"
+        echo "$(md5sum "$input_file" | awk '{ print $1 }')" > "$md5_file"
     fi
-    mkdir -p "$TRANSCODE_DB/$(dirname "$input_file")"
-    echo "$(md5sum "$input_file" | awk '{ print $1 }')" > "$md5_file"
 }
 
 directory_structure()
 {
-    rsync -rvz --exclude-from="./transcode_exclude.cfg" --include="*/" --exclude="*" $TRANSCODE_INPUT_DIR $TRANSCODE_OUTPUT_DIR
+    if [ $MODE_DRY_RUN == true ]; then
+        DRYRUN_FLAG="--dry-run"
+    else
+        DRYRUN_FLAG=""
+    fi
+    echo ""
+    echo "INFO: Creating directory structure with rsync..."
+    rsync -rvz $DRYRUN_FLAG --exclude-from="./transcode_exclude.cfg" --include="*/" --exclude="*" $TRANSCODE_INPUT_DIR $TRANSCODE_OUTPUT_DIR
 }
 
 convert_covers()
-{
+{   
+    echo "INFO: Looking for covers to convert..."
     cd $TRANSCODE_INPUT_DIR
 
     trap "exit" INT
@@ -157,13 +169,13 @@ convert_covers()
         for val in "${StringArray[@]}"; do
             if [ ! -z "$val" ]; then
                 FILENAME="$TRANSCODE_OUTPUT_DIR$(dirname "$val")/$(basename "$val" .$ext).jpg"
-                MD5_FILENAME="$TRANSCODE_DB/$(dirname "$val")/$(basename "$val" .$ext).jpg.md5"
+                MD5_FILENAME="$TRANSCODE_DB/$(dirname "$val")/$(basename "$val").md5"
                 # Check if a MD5 checksum already exists
                 RESULT=$($TRANSCODE_FD_BIN . "$(dirname "$MD5_FILENAME")" | grep -F "$(basename "$MD5_FILENAME")")
                 if [ $? -ne 0 ] ; then
                     write_jpg "$val" "$FILENAME" "$MD5_FILENAME"
                 # Check if an existing MD5 checksum is different
-                elif [ MODE_CHECKSUM == true ]; then
+                elif [ $MODE_CHECKSUM == true ]; then
                     if [ "$(cat "$MD5_FILENAME")" != "$(md5sum "$val" | awk '{ print $1 }')" ]; then
                         write_jpg "$val" "$FILENAME" "$MD5_FILENAME"
                     fi
@@ -175,6 +187,7 @@ convert_covers()
 
 convert_music()
 {
+    echo "INFO: Looking for music to transcode..."
     cd $TRANSCODE_INPUT_DIR
 
     trap "exit" INT
@@ -192,7 +205,7 @@ convert_music()
                 if [ $? -ne 0 ] ; then
                     transcode "$val" "$FILENAME" "$MD5_FILENAME"
                 # Check if an existing MD5 checksum is different
-                elif [ MODE_CHECKSUM == true ]; then
+                elif [ $MODE_CHECKSUM == true ]; then
                     if [ "$(cat "$MD5_FILENAME")" != "$(md5sum "$val" | awk '{ print $1 }')" ]; then
                         transcode "$val" "$FILENAME" "$MD5_FILENAME"
                     fi
@@ -204,6 +217,7 @@ convert_music()
 
 fix_cuefiles()
 {
+    echo "INFO: Looking for cuefiles..."
     cd $TRANSCODE_INPUT_DIR
 
     FILES=$($TRANSCODE_FD_BIN --extension cue $TRANSCODE_FD_FILTERS)
@@ -215,11 +229,13 @@ fix_cuefiles()
             # Check if a MD5 checksum already exists
             RESULT=$($TRANSCODE_FD_BIN . "$(dirname "$MD5_FILENAME")" | grep -F "$(basename "$MD5_FILENAME")")
             if [ $? -ne 0 ] ; then
-                write_cue "$val" "$REPLACEMENT_TEXT_STRING" "$MD5_FILENAME"
+                cp -pr "$val" "$TRANSCODE_OUTPUT_DIR/$val"
+                write_cue "$val" "$TRANSCODE_OUTPUT_DIR/$val" "$REPLACEMENT_TEXT_STRING" "$MD5_FILENAME"
             # Check if an existing MD5 checksum is different
-            elif [ MODE_CHECKSUM == true ]; then
+            elif [ $MODE_CHECKSUM == true ]; then
                 if [ "$(cat "$MD5_FILENAME")" != "$(md5sum "$val" | awk '{ print $1 }')" ]; then
-                    write_cue "$val" "$REPLACEMENT_TEXT_STRING" "$MD5_FILENAME"
+                cp -pr "$val" "$TRANSCODE_OUTPUT_DIR/$val"
+                    write_cue "$val" "$TRANSCODE_OUTPUT_DIR/$val" "$REPLACEMENT_TEXT_STRING" "$MD5_FILENAME"
                 fi
             fi
         fi
@@ -238,24 +254,33 @@ remove_absent_from_source()
         for val in "${StringArray[@]}"; do
             if [ ! -z "$val" ]; then
                 FILENAME=$(dirname "$val")/$(basename "$val" .$ext)
-                RESULT=$($TRANSCODE_FD_BIN . "$TRANSCODE_INPUT_DIR/$(dirname "$FILENAME")" | grep -F "$(basename "$FILENAME")")
+                RESULT=$($TRANSCODE_FD_BIN . "$TRANSCODE_INPUT_DIR/$(dirname "$FILENAME")" | grep -F "$(basename "$FILENAME" .$ext)")
                 # Transcoded file don't have a source file : delete
                 if [ $? -ne 0 ]; then
                     echo "INFO: Transcoded file $FILENAME don't have a source file : delete"
-                    rm "$TRANSCODE_OUTPUT_DIR/$FILENAME"*
-                    rm "$TRANSCODE_DB/$FILENAME"*
+                    if [ $MODE_DRY_RUN == false ]; then
+                        rm "$TRANSCODE_OUTPUT_DIR/$FILENAME"*
+                        rm "$TRANSCODE_DB/$FILENAME"*
+                    fi
                 fi
-                
             fi
         done
     done
 
     echo "INFO: removing empty directories..."
-    cd $TRANSCODE_OUTPUT_DIR
-    find . -type d -empty -delete
+    
+    if [ $MODE_DRY_RUN == false ]; then
+        DIRECTORIES="$TRANSCODE_OUTPUT_DIR $TRANSCODE_DB"
+        mapfile -t StringArray <<< "$DIRECTORIES"
+        for val in "${StringArray[@]}"; do
+            cd $val
+                fd --type empty --exec-batch rmdir
+        done
+    fi
 }
 
-init
+cp -r ./transcode_exclude.cfg $TRANSCODE_INPUT_DIR/.fdignore
+cp -r ./transcode_exclude.cfg $TRANSCODE_OUTPUT_DIR/.fdignore
 
 if [ $MODE_DELETE == false ]; then
     directory_structure
@@ -269,4 +294,5 @@ else
     remove_absent_from_source
 fi
 
-clean
+rm $TRANSCODE_INPUT_DIR/.fdignore
+rm $TRANSCODE_OUTPUT_DIR/.fdignore
